Update example_weighted.sh
make clustering possible with the example data.
Previous example had the parameters make classification impossible (since the maximal weight 150 was too constraining for the dataset with weight sum around 7,000) and also mismatch with the comment text which stated maximal weight 50.
Changed it to include two examples. One almost doesn't constrain the minimal weight (min_weight=1) and one constrains it to 2,000 so the resulting clustering is different.